requirements.txt
Django==5.0.1
django-allauth==0.60.1
django-allauth-ui==0.2.2
django-widget-tweaks==1.5.0
